backend/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
numpy
scikit-learn
cohere
python-dotenv
